# api/swagger.yaml
swagger: "2.0"
info:
  description: "This is the cloud-runner API definition"
  version: "0.1.0"
  title: "Cloud Runner"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "cloud-runner.cd:80"
...
schemes:
- "https"
- "http"
paths:
...
  /v1/credentials:
    get:
      tags:
      - "credentials"
      summary: "List credentials"
      produces:
      - "application/json"
      parameters:
      - name: "onlyForUser"
        in: "query"
        description: "If true, only retrieve the accounts for the auth'd user"
        required: false
        type: "string"
      - name: "X-Spinnaker-User"
        in: "header"
        description: "The user making the request"
        required: false
        type: "string"
      responses:
        "200":
          description: "List of credentials"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Credentials"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: "Generic error"
          schema:
            $ref: "#/definitions/Error"
    post:
      tags:
      - "credentials"
      summary: "Create credentials for a given GCP project"
      description: ""
      operationId: "createCredentials"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Object that describes the credentials"
        required: true
        schema:
          $ref: "#/definitions/Credentials"
      - in: "header"
        name: "apiKey"
        description: "API Key required for create operations"
        type: "string"
        required: true
      responses:
        "201":
          description: "Credentials created"
          schema:
            $ref: "#/definitions/Credentials"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        "409":
          description: "Credentials already exist"
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: "Generic error"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Credentials:
    type: "object"
    properties:
      account:
        type: "string"
        example: "cr-test-project-id-pr"
      lifecycle:
        type: "string"
        example: "pr"
      projectID:
        type: "string"
        example: "test-project-id"
      readGroups:
        type: "array"
        items:
          type: "string"
          example: "gg_my_group"
      writeGroups:
        type: "array"
        items:
          type: "string"
          example: "gg_my_group"
...
  Error:
    type: "object"
    properties:
      error:
        type: "string"
        example: "an error occurred"
...